Move showUpdateAlert method in GameUtils.java to UpdateChecker.java
